+++ documentation/docs/release-notes.md
@@ -7,6 +7,7 @@ title: Release notes&#58;
 
 **v6.2.1**:
 - Fix the SAML `IssueInstant` check
+- Fix the conditional `Authorization` header
 
 **v6.2.0**:
 - Only return an `Authorization` header when there is no credentials or when credentials are invalid (for `DirectBasicAuthClient`, `DirectBearerAuthClient` and `DirectDigestAuthClient`)